Fixed Observer problem
When creating a observer / subject then observers get registered in a raw pointer vector in the subject.
The problem is that if you forget calling Unregister function you can have big runtime errors.
So to avoid this I made sure that every observer adds and removes itself from the subject.